requirements.txt
Django==2.0.3
django-configurations==2.0
django-extensions==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7